chore: batch dependabot dependency updates (#280)
* chore(deps): Bump rusqlite from 0.34.0 to 0.40.0

Bumps [rusqlite](https://github.com/rusqlite/rusqlite) from 0.34.0 to 0.40.0.
- [Release notes](https://github.com/rusqlite/rusqlite/releases)
- [Changelog](https://github.com/rusqlite/rusqlite/blob/master/Changelog.md)
- [Commits](https://github.com/rusqlite/rusqlite/compare/v0.34.0...v0.40.0)

---
updated-dependencies:
- dependency-name: rusqlite
  dependency-version: 0.40.0
  dependency-type: direct:production
  update-type: version-update:semver-minor
...


* chore(ci): Bump actions/checkout in the actions group

Bumps the actions group with 1 update: [actions/checkout](https://github.com/actions/checkout).


Updates `actions/checkout` from 6.0.2 to 6.0.3
- [Release notes](https://github.com/actions/checkout/releases)
- [Changelog](https://github.com/actions/checkout/blob/main/CHANGELOG.md)
- [Commits](https://github.com/actions/checkout/compare/de0fac2e4500dabe0009e67214ff5f5447ce83dd...df4cb1c069e1874edd31b4311f1884172cec0e10)

---
updated-dependencies:
- dependency-name: actions/checkout
  dependency-version: 6.0.3
  dependency-type: direct:production
  update-type: version-update:semver-patch
  dependency-group: actions
...


* chore(deps): Bump the rust-dependencies group across 1 directory with 4 updates

Bumps the rust-dependencies group with 4 updates in the / directory: [chrono](https://github.com/chronotope/chrono), [uuid](https://github.com/uuid-rs/uuid), [http](https://github.com/hyperium/http) and [libz-sys](https://github.com/rust-lang/libz-sys).


Updates `chrono` from 0.4.44 to 0.4.45
- [Release notes](https://github.com/chronotope/chrono/releases)
- [Changelog](https://github.com/chronotope/chrono/blob/main/CHANGELOG.md)
- [Commits](https://github.com/chronotope/chrono/compare/v0.4.44...v0.4.45)

Updates `uuid` from 1.23.1 to 1.23.3
- [Release notes](https://github.com/uuid-rs/uuid/releases)
- [Commits](https://github.com/uuid-rs/uuid/compare/v1.23.1...v1.23.3)

Updates `http` from 1.4.1 to 1.4.2
- [Release notes](https://github.com/hyperium/http/releases)
- [Changelog](https://github.com/hyperium/http/blob/master/CHANGELOG.md)
- [Commits](https://github.com/hyperium/http/compare/v1.4.1...v1.4.2)

Updates `libz-sys` from 1.1.28 to 1.1.29
- [Release notes](https://github.com/rust-lang/libz-sys/releases)
- [Commits](https://github.com/rust-lang/libz-sys/compare/1.1.28...1.1.29)

---
updated-dependencies:
- dependency-name: chrono
  dependency-version: 0.4.45
  dependency-type: direct:production
  update-type: version-update:semver-patch
  dependency-group: rust-dependencies
- dependency-name: uuid
  dependency-version: 1.23.3
  dependency-type: direct:production
  update-type: version-update:semver-patch
  dependency-group: rust-dependencies
- dependency-name: http
  dependency-version: 1.4.2
  dependency-type: direct:production
  update-type: version-update:semver-patch
  dependency-group: rust-dependencies
- dependency-name: libz-sys
  dependency-version: 1.1.29
  dependency-type: direct:production
  update-type: version-update:semver-patch
  dependency-group: rust-dependencies
...


* chore: batch dependabot dependency updates

Merge all open dependabot PRs into a single update:

- actions/checkout 6.0.2 -> 6.0.3 (#272)
- chrono 0.4.44 -> 0.4.45, uuid 1.23.1 -> 1.23.3,
  http 1.4.1 -> 1.4.2, libz-sys 1.1.28 -> 1.1.29 (#275)
- rusqlite 0.34.0 -> 0.40.0 (#266)

Fix rusqlite 0.40 breaking changes: usize no longer implements
ToSql, and u64 no longer implements FromSql. Cast through i64
in recall_log.rs.


* fix: pin rusqlite to 0.39 to restore MSRV 1.88 compatibility

rusqlite 0.40 pulls in libsqlite3-sys 0.38 which uses cfg_select!,
a macro not available until Rust 1.90+. Downgrade to rusqlite 0.39
(libsqlite3-sys 0.37) which compiles cleanly on the 1.88 MSRV.


* chore: bump MSRV to 1.95, unpin rusqlite to 0.40

The previous commit pinned rusqlite to 0.39 as a workaround for the
MSRV constraint. With MSRV raised to 1.95, rusqlite 0.40 compiles
cleanly — unpin it and update the lock file accordingly.


---------

+++ src/recall_log.rs
@@ -122,7 +122,7 @@ impl RecallLog {
                     session_id,
                     r.memory_name,
                     r.scope,
-                    r.rank,
+                    r.rank as i64,
                     r.distance
                 ])
                 .map_err(|e| MemoryError::Internal(format!("recall log insert: {e}")))?;
@@ -274,11 +274,11 @@ impl RecallLog {
             .query_map([], |row| {
                 Ok((
                     row.get::<_, i64>(0)? as usize,
-                    row.get::<_, u64>(1)?,
-                    row.get::<_, u64>(2)?,
-                    row.get::<_, u64>(3)?,
-                    row.get::<_, u64>(4)?,
-                    row.get::<_, u64>(5)?,
+                    row.get::<_, i64>(1)? as u64,
+                    row.get::<_, i64>(2)? as u64,
+                    row.get::<_, i64>(3)? as u64,
+                    row.get::<_, i64>(4)? as u64,
+                    row.get::<_, i64>(5)? as u64,
                 ))
             })
             .map_err(|e| MemoryError::Internal(format!("recall_stats query: {e}")))?;
